pyproject.toml
[project]
name = "youtube-data-api-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.169.0",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.15.0",
    "ruff>=0.11.8",
]
